refactor(quotes): move trade sampler to common package
