+++ pkg/quotes/common/trade_sampler.go
@@ -1,4 +1,4 @@
-package quotes
+package common
 
 import (
 	"math/rand"
